Add typedef. Fix build.
